add form tests refactoring
